fix index type (str->int)

+++ deeptools/computeGCBias_readlen.py
@@ -471,9 +471,8 @@ def interpolate_ratio(df):
     F_GC = df.loc["F_gc_reads"]
     
     # get min and max values
-    N_GC_min, N_GC_max =  np.nanmin(N_GC.index), np.nanmax(N_GC.index)
-    F_GC_min, F_GC_max =  np.nanmin(F_GC.index), np.nanmax(F_GC.index)
-    
+    N_GC_min, N_GC_max =  np.nanmin(N_GC.index.astype("int")), np.nanmax(N_GC.index.astype("int"))
+    F_GC_min, F_GC_max =  np.nanmin(F_GC.index.astype("int")), np.nanmax(F_GC.index.astype("int"))
     # sparse grid for hypothetical read density
     N_GC_readlen = N_GC.index.to_numpy(dtype=int)
     N_GC_gc = N_GC.columns.to_numpy(dtype=int)
